[refactor] coverLetter types > shared로 통합
