--- frontend/src/features/coverLetter/types/coverLetter.ts
@@ -1,24 +1,0 @@
-export interface ScrapItem {
-  id: number;
-  companyName: string;
-  jobPosition: string;
-  applySeason: string; // "2025년 하반기"
-  question: string;
-  answer: string;
-  coverLetterId: number;
-}
-
-export interface GetScrapsResponse {
-  scraps: ScrapItem[];
-  hasNext: boolean;
-}
-
-export interface ErrorResponse {
-  message: string;
-}
-
-export interface TextChangeResult {
-  changeStart: number;
-  oldLength: number;
-  newLength: number;
-}
