+++ frontend/src/shared/types/coverLetter.ts
@@ -64,3 +64,28 @@ export interface FilterRequest {
   size?: number;
   isShared?: boolean;
 }
+
+export interface ScrapItem {
+  id: number;
+  companyName: string;
+  jobPosition: string;
+  applySeason: string; // "2025년 하반기"
+  question: string;
+  answer: string;
+  coverLetterId: number;
+}
+
+export interface GetScrapsResponse {
+  scraps: ScrapItem[];
+  hasNext: boolean;
+}
+
+export interface ErrorResponse {
+  message: string;
+}
+
+export interface TextChangeResult {
+  changeStart: number;
+  oldLength: number;
+  newLength: number;
+}
